Correct pointer path
The old pointer path points to a `level` variable in-memory, which updates about half a second after the button was pressed. Moreover, the other offsets for the final IGT used in the autosplitter to set the final time is not available for the old pointer path.
This should now update to the correct pointer path.