Added multiple features
Added splitting support for buttons that do not raise new platforms
Added splitting support for cubes
Added settings for every individual button
Added a setting to allow users to start the timer when unpausing

Please for the love of god clean up this code. This is the best solution I could come up with.